+++ packages/plugins/graphql-middleware/src/index.ts
@@ -1,20 +1,24 @@
 import { Plugin } from '@envelop/core';
 import { IMiddlewareGenerator, IMiddleware, applyMiddleware } from 'graphql-middleware';
 
-const middlewaresAppliedSymbol = Symbol('SCHEMA_WITH_MIDDLEWARES');
+const graphqlMiddlewareAppliedTransformSymbol = Symbol('graphqlMiddleware.appliedTransform');
 
 export const useGraphQLMiddleware = <TSource = any, TContext = any, TArgs = any>(
   middlewares: (IMiddleware<TSource, TContext, TArgs> | IMiddlewareGenerator<TSource, TContext, TArgs>)[]
 ): Plugin => {
   return {
     onSchemaChange({ schema, replaceSchema }) {
-      if (schema[middlewaresAppliedSymbol]) {
+      // @ts-expect-error See https://github.com/graphql/graphql-js/pull/3511 - remove this comments once merged
+      if (schema.extensions?.[graphqlMiddlewareAppliedTransformSymbol]) {
         return;
       }
 
       if (middlewares.length > 0) {
         const wrappedSchema = applyMiddleware(schema, ...middlewares);
-        wrappedSchema[middlewaresAppliedSymbol] = true;
+        wrappedSchema.extensions = {
+          ...schema.extensions,
+          [graphqlMiddlewareAppliedTransformSymbol]: true,
+        };
         replaceSchema(wrappedSchema);
       }
     },