+++ packages/plugins/execute-subscription-event/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @envelop/execute-subscription-event
 
+## 2.3.2
+
+### Patch Changes
+
+- Updated dependencies [07d029b]
+  - @envelop/core@2.3.2
+
 ## 2.3.1
 
 ### Patch Changes