+++ packages/plugins/prometheus/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @envelop/prometheus
 
+## 6.3.2
+
+### Patch Changes
+
+- Updated dependencies [07d029b]
+  - @envelop/core@2.3.2
+
 ## 6.3.1
 
 ### Patch Changes